+++ src/common/Types.ts
@@ -3,7 +3,16 @@
  * @license MIT
  */
 
-import { IEventEmitter } from 'xterm';
+export interface IDisposable {
+  dispose(): void;
+}
+
+export interface IEventEmitter {
+  on(type: string, listener: (...args: any[]) => void): void;
+  off(type: string, listener: (...args: any[]) => void): void;
+  emit(type: string, data?: any): void;
+  addDisposableListener(type: string, handler: (...args: any[]) => void): IDisposable;
+}
 
 export type XtermListener = (...args: any[]) => void;
